requirements.txt
amqp==5.2.0
asgiref==3.8.1
autopep8==2.3.1
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.16
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-silk==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docopt==0.6.2
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.10.2
geventhttpclient==2.3.1
gprof2dot==2024.6.6
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
locust==2.31.8
MarkupSafe==3.0.1
model-bakery==1.20.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0
